extrapolated key POST functions from SurveyContainer.
